# perf(linter/numeric-separators-style): replace regex with number parser